Cuando se mueve el slider de cambiar la velocidad del agente físico. Se envía a través del socket el mensaje para que se cambie.

+++ src/Networking/Conexion_SMA.java
@@ -83,7 +83,7 @@ public class Conexion_SMA extends DataServer
     }
     
     
-    public void enviarNuevaVelocidad(int v)
+    public void enviarNuevaVelocidad(float v)
     {
         String sms = Encabezado_Mensajes.Msj_ACOtoSMA_setVelocidad + Encabezado_Mensajes.Msj_divisor + v;
